refactor: update loading, move hooks related to auth
