+++ frontend/src/hooks/auth/useSignIn.hook.ts
@@ -1,17 +1,23 @@
 import { useState } from "react";
-import { useAuthContext } from "../context/AuthContext";
+import { useAuthContext } from "../../context/AuthContext";
 import toast from "react-hot-toast";
 
-export const useLogout = () => {
+type SignInInputs = {
+  username: string;
+  password: string;
+};
+
+export const useSignIn = () => {
   const [loading, setLoading] = useState(false);
   const { setAuthUser } = useAuthContext();
 
-  const logout = async () => {
+  const signin = async (inputs: SignInInputs) => {
     setLoading(true);
     try {
-      const res = await fetch("/api/auth/logout", {
+      const res = await fetch("/api/auth/signin", {
         method: "POST",
         headers: { "Content-Type": "application/json" },
+        body: JSON.stringify(inputs),
       });
 
       if (!res.ok) {
@@ -19,8 +25,9 @@ export const useLogout = () => {
         throw new Error(error);
       }
 
-      setAuthUser(null);
-      toast.success("Logged out successfully!");
+      const data = await res.json();
+      setAuthUser(data);
+      toast.success("Signed in successfully!");
     } catch (error: any) {
       toast.error(error.message);
       throw error;
@@ -29,5 +36,5 @@ export const useLogout = () => {
     }
   };
 
-  return { loading, logout };
+  return { loading, signin };
 };
