+++ frontend/src/hooks/auth/useSignInForm.hook.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useSignIn } from "./hooks";
+import { useSignIn } from "../hooks";
 import toast from "react-hot-toast";
 
 export const useSignInForm = () => {
